// generated/src/app/api-models/auth-provider.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class AuthProvider {
  authNumber: number;
  secondaryAuthNo: string;
  provCode: string;
  seqProvId: number;
  prevSeqProvId: number;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  seqProvContract: number;
  outOfNetStatus: string;
  seqProvxId: number;
  seqCovProvGrp: number;
  servicePostalCode: string;
  inAreaServiceStatus: string;
  provTaxonomyCode: string;
  seqVendId: number;
}